Improve reconnection support in Connection\Manager
When switching nodes, the prioritizer can recommend to stay on the current node.
The current node however could have lost its connection in the meantime.
In that case we will try and reconnect, or if not possible, connect to the next
best possible node.

The check whether the current node has disconnected has been moved so that
it is checked every time we execute any command, not just after the GETJOB
command.

If called multiple times, the method Manager::connect() will first try to
reconnect based on the latest cluster information.

If the reconnection attempt doesn't succeed, or if it's the first call,
it will try to connect based on the user-supplied credentials.

Also if the Manager reconnects to a disconnected node that was used previously,
don't lose the node stats.