Reconnect, if the current node has lost its connection
When switching nodes, the prioritizer can recommend to stay on the current node.
The current node however could have lost its connection in the meantime.
In that case we will try and reconnect, or if not possible, connect to the next
best possible node.

The check whether the current node has disconnected has been moved so that
it is checked every time we execute any command, not just after the GETJOB
command.